+++ lib/honeybadger/logger.ex
@@ -1,30 +1,56 @@
 defmodule Honeybadger.Logger do
-  alias Honeybadger.Utils
+  @moduledoc false
+
+  @behaviour :gen_event
 
   require Logger
 
-  use GenEvent
+  alias Honeybadger.Utils
 
-  def init(_mod, []), do: {:ok, []}
+  def init(args) do
+    {:ok, args}
+  end
+
+  ## Callbacks
+
+  def handle_event({_type, gl, _msg}, state) when node(gl) != node() do
+    {:ok, state}
+  end
+
+  def handle_event(event, state) do
+    handle_error(event)
+
+    {:ok, state}
+  end
 
   def handle_call({:configure, new_keys}, _state) do
     {:ok, :ok, new_keys}
   end
 
-  def handle_event({_level, gl, _event}, state)
-      when node(gl) != node() do
+  def handle_call(request, _state) do
+    exit {:bad_call, request}
+  end
+
+  def handle_info(_msg, state) do
     {:ok, state}
   end
 
-  def handle_event({:error_report, _gl, {_pid, _type, [message | _]}}, state)
-      when is_list(message) do
-    try do
-      error_info = message[:error_info]
-      context = get_in(message, [:dictionary, :honeybadger_context])
-      context = %{context: context}
+  def code_change(_old_vsn, state, _extra) do
+    {:ok, state}
+  end
 
-      case error_info do
-        {_kind, {exception, stacktrace}, _stack} when is_list(stacktrace) ->
+  def terminate(_reason, _state) do
+    :ok
+  end
+
+  ## Helpers
+
+  defp handle_error({:error_report, _gl, {_pid, _type, [message | _]}}) when is_list(message) do
+    try do
+      context = get_in(message, [:dictionary, :honeybadger_context])
+
+      case message[:error_info] do
+        {_kind, {exception, stacktrace}, _stack} ->
           Honeybadger.notify(exception, context, stacktrace)
         {_kind, exception, stacktrace} ->
           Honeybadger.notify(exception, context, stacktrace)
@@ -38,11 +64,9 @@ defmodule Honeybadger.Logger do
           "Unable to notify Honeybadger! #{error_type}: #{reason}"
         end)
     end
-
-    {:ok, state}
   end
 
-  def handle_event({_level, _gl, _event}, state) do
-    {:ok, state}
+  defp handle_error(_event) do
+    :ok
   end
 end